fix dm collectors and add reaction role

+++ src/util/validate.js
@@ -15,15 +15,22 @@ module.exports.validateOptions = (options, type) => {
         throw 'Invalid input: botMessage is undefined or invalid.';
 
     const client = options.botMessage.client;
-    options.user = client.users.resolve(options.user);
+    if (typeof client.users.resolve == 'function')
+        options.user = client.users.resolve(options.user);
+    else
+        options.user = client.resolver.resolveUser(options.user);
     if (!options.user)
         throw 'Invalid input: user is undefined or invalid.';
-    if (!options.botMessage.guild)
-        throw 'Invalid input: botMessage.guild is undefined.';
-    if (!options.botMessage.guild || !options.botMessage.guild.me)
-        throw 'Invalid input: botMessage.guild.me is undefined.';
-    if (!options.botMessage.guild.me.permissionsIn(options.botMessage.channel).has('ADD_REACTIONS'))
-        throw 'Missing permissions: I cannot react in messages in that channel.';
+
+    if (options.botMessage.channel.type == 'text') {
+        if (!options.botMessage.guild)
+            throw 'Invalid input: botMessage.guild is undefined.';
+
+        if (options.botMessage.guild.me) {
+            if (!options.botMessage.guild.me.permissionsIn(options.botMessage.channel).has('ADD_REACTIONS'))
+                throw 'Missing permissions: I cannot react in messages in that channel.';
+        }
+    }
 
     switch (type) {
         case 'reactQuestion':
@@ -36,26 +43,34 @@ module.exports.validateOptions = (options, type) => {
 
             if (!options.reactions)
                 options.reactions = (type !== 'reactPaginator') ? Constants.DEFAULT_YES_NO_REACTIONS : Constants.DEFAULT_PAGINATOR_REACTIONS;
-            
+
             if (!options.onReact)
                 options.onReact = [Constants.DEFAULT_RETURN_FUNCTION, Constants.DEFAULT_RETURN_FUNCTION];
 
-            if (options.deleteReaction === undefined)
+            if (options.botMessage.channel.type == 'dm')
+                options.deleteReaction = false;
+            else if (options.deleteReaction === undefined)
                 options.deleteReaction = true;
             else if (!isBoolean(options.deleteReaction))
                 options.deleteReaction = Boolean(options.deleteReaction);
 
-            if (options.deleteAllReactionsWhenCollectorEnd === undefined)
+            if (options.botMessage.channel.type == 'dm')
+                options.deleteAllReactionsWhenCollectorEnd = false;
+            else if (options.deleteAllReactionsWhenCollectorEnd === undefined)
                 options.deleteAllReactionsWhenCollectorEnd = true;
             else if (!isBoolean(options.deleteAllReactionsWhenCollectorEnd))
                 options.deleteAllReactionsWhenCollectorEnd = Boolean(options.deleteAllReactionsWhenCollectorEnd);
 
-            if (options.deleteReaction && !options.botMessage.guild.me.permissionsIn(options.botMessage.channel).has('MANAGE_MESSAGES'))
-                throw 'Missing permissions: I not have permissions to Manage Messages in this channel to delete reactions.';
+            if (options.botMessage.channel.type == 'text') {
+                if (!options.botMessage.guild.me.permissionsIn(options.botMessage.channel).has('ADD_REACTIONS'))
+                    throw 'Missing permissions: I cannot react in messages in that channel.';
 
-            if (options.deleteAllReactionsWhenCollectorEnd && !options.botMessage.guild.me.permissionsIn(options.botMessage.channel).has('MANAGE_MESSAGES'))
-                throw 'Missing permissions: I not have permissions to Manage Messages in this channel to delete all reactions when collector end.';
+                if (options.deleteReaction && !options.botMessage.guild.me.permissionsIn(options.botMessage.channel).has('MANAGE_MESSAGES'))
+                    throw 'Missing permissions: I not have permissions to Manage Messages in this channel to delete reactions.';
 
+                if (options.deleteAllReactionsWhenCollectorEnd && !options.botMessage.guild.me.permissionsIn(options.botMessage.channel).has('MANAGE_MESSAGES'))
+                    throw 'Missing permissions: I not have permissions to Manage Messages in this channel to delete all reactions when collector end.';
+            }
             break;
 
         case 'messageQuestion':
@@ -65,25 +80,27 @@ module.exports.validateOptions = (options, type) => {
 
             if (!options.onMessage)
                 options.onMessage = Constants.DEFAULT_RETURN_FUNCTION;
-            
-            if (!isBoolean(options.deleteMessage))
+
+            if (options.botMessage.channel.type == 'dm')
+                options.deleteMessage = false;
+            else if (!isBoolean(options.deleteMessage))
                 options.deleteMessage = Boolean(options.deleteMessage);
 
-            if (options.deleteMessage && !options.botMessage.guild.me.permissionsIn(options.botMessage.channel).has('MANAGE_MESSAGES'))
+            if (options.botMessage.channel.type == 'text' && options.deleteMessage && !options.botMessage.guild.me.permissionsIn(options.botMessage.channel).has('MANAGE_MESSAGES'))
                 throw 'Missing permissions: I not have permissions to Manage Messages in this channel to delete messages.';
             break;
     }
 
     if (!options.collectorOptions || !isObject(options.collectorOptions))
         options.collectorOptions = { time: Constants.DEFAULT_COLLECTOR_TIME, max: (type === 'reactPaginator') ? Constants.DEFAULT_PAGINATOR_MAX_REACT : Constants.DEFAULT_COLLECTOR_MAX_REACT };
-        
+
     if (!isNumber(options.collectorOptions.time)) {
         options.collectorOptions.time = parseInt(options.collectorOptions.time);
         if (isNaN(options.collectorOptions.time)) {
             options.collectorOptions.time = Constants.DEFAULT_COLLECTOR_TIME;
         }
     }
-        
+
     if (!isNumber(options.collectorOptions.max)) {
         options.collectorOptions.max = parseInt(options.collectorOptions.max);
         if (isNaN(options.collectorOptions.max)) {
